Fix non-jpegli CodeQL findings (#4422)
(cherry picked from commit 1bb7f02535a006afbc753282f37a7ec348afe3d1)

+++ lib/extras/metrics.cc
@@ -212,7 +212,7 @@ Status ComputeSumOfSquares(JxlMemoryManager* memory_manager,
         }
       }
       for (size_t j = 0; j < 3; j++) {
-        sum_of_squares[j] += yuvdiff[j] * yuvdiff[j];
+        sum_of_squares[j] += static_cast<double>(yuvdiff[j]) * yuvdiff[j];
       }
     }
   }
